# [ArrowStringArray] PERF: bypass some padding code in _wrap_result